Move `requires_compiler` under `tapioca::static`
