+++ lib/tapioca/internal.rb
@@ -15,10 +15,10 @@ require "tapioca/generators"
 require "tapioca/cli"
 require "tapioca/gemfile"
 require "tapioca/executor"
+require "tapioca/static/requires_compiler"
 require "tapioca/static/symbol_table_parser"
 require "tapioca/static/symbol_loader"
 require "tapioca/compilers/gem/events"
 require "tapioca/compilers/gem/listeners"
-require "tapioca/compilers/requires_compiler"
 require "tapioca/compilers/symbol_table_compiler"
 require "tapioca/dsl/pipeline"
